Renamed handlers/widgets.py -> handlers/dashboard_widgets.py
