+++ redash/handlers/dashboard_widgets.py
