+++ redash/handlers/embed.py
@@ -31,7 +31,10 @@ def embed(query_id, visualization_id, org_slug=None):
     return render_index()
 
 
-@routes.route(org_scoped_rule("/public/dashboards/<token>"), methods=["GET"])
+@routes.route(
+    org_scoped_rule("/public/dashboards/<token>"),
+    methods=["GET"],
+)
 @login_required
 @csp_allows_embeding
 def public_dashboard(token, org_slug=None):
